requirements.txt
#  Core Langchain libraries
langchain
langchain-core
langchain-community
langchain-classic
langchain_text_splitters
langchain-experimental
langchain-pinecone

# Google Generative AI Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# ML Utilities
numpy
scikit-learn
torch
sentence-transformers
streamlit
pydantic
datasets
requests

# Visual-Graph
grandalf

# PDF
pypdf

# URL Scraper
beautifulsoup4

# Vector Database
pinecone
pinecone-client